src/backend/rag/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.24.3 